Compleate Compleate

+++ Controller/contact.Controller.js
@@ -3,15 +3,15 @@ const contactModel = require("../models/contact.Model")
 const { ObjectId } = require('mongodb');
 //@dec get all contact
 //@route get /book
-//@access public
+//@access Private
 const getAllContact = asyncHandler(async (req, res) => {
-    const connect = await contactModel.find();
+    const connect = await contactModel.find({user_id : req.user.id});
     res.status(200).json(connect);
 })
 
 //@dec get songle contact
 //@route get /book/id
-//@access public
+//@access Private
 const getSingleContact =asyncHandler(async (req, res) => {
     const contact = await contactModel.findById(req.params.id);
     if(!contact){
@@ -23,7 +23,7 @@ const getSingleContact =asyncHandler(async (req, res) => {
 
 //@dec get add contact
 //@route Post /book
-//@access public
+//@access Private
 const addContact =asyncHandler( async (req, res) => {
     console.log("the request Body is ", req.body);
     const { name, email, phone } = req.body;
@@ -34,14 +34,15 @@ const addContact =asyncHandler( async (req, res) => {
     const contact = await contactModel.create({
         name,
         email,
-        phone
+        phone,
+        user_id :req.user.id
     })
     res.status(201).json(contact);
 })
 
 //@dec get Put contact
 //@route Put /book/id
-//@access public
+//@access Private
 const UpdateContact = asyncHandler(async (req, res) => {
     const body = req.body;
     const id = new ObjectId(req.params.id);
@@ -51,13 +52,18 @@ const UpdateContact = asyncHandler(async (req, res) => {
         res.status(404);
         throw new Error("contact not found")
     }
+
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("user dont have permission to update the contact")
+    }
     const updateContact = await contactModel.findByIdAndUpdate(id,body,option);
     res.status(200).json(updateContact);
 })
 
 //@dec get Delete contact
 //@route delete /book/id
-//@access public
+//@access Private
 const detleteContact =asyncHandler( async (req, res) => {
     const contact = await contactModel.findById(new ObjectId(req.params.id));
     if(!contact){
@@ -65,6 +71,10 @@ const detleteContact =asyncHandler( async (req, res) => {
         throw new Error("contact not found")
     }
     // await contactModel.deleteById(new ObjectId(req.params.id))
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("user dont have permission to update the contact")
+    }
     await contactModel.deleteOne(new ObjectId(req.params.id))
